requirements.txt
Flask==2.2.3
pymysql==1.0.2
Werkzeug==2.2.3
Flask-Cors==3.0.10
PyJWT==2.6.0